add items controller
