+++ app/controllers/yablog/admin/content_statuses_controller.rb
@@ -1,5 +1,5 @@
 module Yablog
-  class Admin::ContentStatusesController < AdminController
+  class Admin::ContentStatusesController < Admin::ItemsController
     MODEL = Yablog::ContentStatus
     ATTRIBUTES = [:title, :description]
     INDEX = [:title, :description]
