+++ app/controllers/yablog/admin/dashboard_controller.rb
@@ -1,4 +1,4 @@
 module Yablog
-  class Admin::DashboardController < ApplicationController
+  class Admin::DashboardController < AdminController
   end
 end
