+++ app/controllers/yablog/admin/pages_controller.rb
@@ -1,5 +1,5 @@
 module Yablog
-  class Admin::PagesController < AdminController
+  class Admin::PagesController < Admin::ItemsController
     MODEL = Yablog::Page
     ATTRIBUTES = [:title, :slug, :description, :body, :tags]
     INDEX = [:title, :description, :body]
