+++ app/controllers/yablog/admin/posts_controller.rb
@@ -1,5 +1,5 @@
 module Yablog
-  class Admin::PostsController < AdminController
+  class Admin::PostsController < Admin::ItemsController
     MODEL = Yablog::Post
     ATTRIBUTES = [:title, :slug, :description, :body, :tags]
     INDEX = [:title, :description, :body, :category]
